Move magma/wire_clock.py methods into magma/passes/clock.py
