+++ tests/test_simulator/test_mdb.py
@@ -5,7 +5,7 @@ from magma.simulator.mdb import SimulationConsole
 from magma import *
 from magma.scope import *
 from magma.passes.debug_name import DebugNamePass
-from magma.wire_clock import drive_undriven_other_clock_types_in_inst
+from magma.passes.clock import drive_undriven_other_clock_types_in_inst
 
 def test(capsys):
     m.config.set_debug_mode(True)
